Add files via upload
I was aiming to learn cybersecurity and its fundamentals with a target deadline of 1wk , and after the day-6 i was happy enough to grasp the 95+% concepts of cybersecurity fundamentals . The notes contains overview of DevSecops  , ethical hacking & pen-testing. If you will follow my roadmap of 6 days and learn everything properly you are going to crack 90% of your screenings and basic interview questions. However, with theories one must have the power to implement all of this in practical real time show that you are ready to handle any situation. At the end i would appreciate your efforts for being with me i am sharing this my journey in digital platform so thar after months or yeaar when i will look back in past i have something beautiful to admire.